+++ bauh/app_args.py
@@ -9,6 +9,8 @@ def read() -> Namespace:
     parser.add_argument('-v', '--version', action='version', version='%(prog)s {}'.format(__version__))
     parser.add_argument('--logs', action="store_true", help='It activates {} logs.'.format(__app_name__))
     parser.add_argument('--offline', action="store_true", help='It assumes the internet connection is off')
+    parser.add_argument('--suggestions', action="store_true",
+                        help='It forces loading software suggestions after the initialization process')
 
     exclusive_args = parser.add_mutually_exclusive_group()
     exclusive_args.add_argument('--tray', action="store_true", help='If {} should be attached to the system tray.'.format(__app_name__))